+++ packages/nx/lib/nx.ml
@@ -74,6 +74,10 @@ let randint ?low ~high shape = F.randint (Lazy.force context) ?low ~high shape
 let bernoulli ~p shape = F.bernoulli (Lazy.force context) ~p shape
 let gumbel dtype shape = F.gumbel (Lazy.force context) dtype shape
 let exponential dtype shape = F.exponential (Lazy.force context) dtype shape
+let gamma ~concentration dtype shape =
+  F.gamma (Lazy.force context) ~concentration dtype shape
+
+let poisson ~rate shape = F.poisson (Lazy.force context) ~rate shape
 let permutation n = F.permutation (Lazy.force context) n
 let shuffle x = F.shuffle (Lazy.force context) x